# photodome-ios/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/events/{eventId}": {
      "get": {
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSnapshotDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Get the authorized event snapshot",
        "tags": [
          "events"
        ]
      }
    },
    "/v1/events/{eventId}/host-transfer": {
      "post": {
        "operationId": "createHostTransfer",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateHostTransferResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Create a short-lived one-time host transfer",
        "tags": [
          "events"
        ]
      }
    },
...
    "/v1/events/{eventId}/end": {
      "post": {
        "operationId": "endEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSnapshotDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "End an event and start its seven-day retention clock",
        "tags": [
          "event controls"
        ]
      }
    },
    "/v1/events/{eventId}/restrict-uploads": {
      "post": {
        "operationId": "restrictEventUploads",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSnapshotDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Block new upload reservations after an event has ended",
        "tags": [
          "event controls"
        ]
      }
    },
    "/v1/events/{eventId}/members": {
      "get": {
        "operationId": "listEventMembers",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventMemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "List active anonymous event members",
        "tags": [
          "event controls"
        ]
      }
    },
    "/v1/events/{eventId}/members/me": {
      "patch": {
        "operationId": "updateOwnEventDisplayName",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDisplayNameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSnapshotDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Update the current member display name",
        "tags": [
          "event controls"
        ]
      }
    },
    "/v1/events/{eventId}/members/{memberId}": {
      "delete": {
        "operationId": "removeEventMember",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Remove an attendee and revoke their event access",
        "tags": [
          "event controls"
        ]
      }
    },
    "/v1/events/{eventId}/rotate-code": {
      "post": {
        "operationId": "rotateEventJoinCode",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RotatedJoinCodeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Rotate the public join code",
        "tags": [
          "event controls"
        ]
      }
    },
...
    "/v1/events/{eventId}/selections/review": {
      "get": {
        "operationId": "getPhotoReviewQueue",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPhotoPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "List this member’s remaining private review queue",
        "tags": [
          "curation"
        ]
      }
    },
    "/v1/events/{eventId}/selections/{photoId}": {
      "put": {
        "operationId": "setPhotoSelection",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPhotoSelectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoSelectionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Privately keep or skip one photo for this member",
        "tags": [
          "curation"
        ]
      }
    },
    "/v1/events/{eventId}/selections/latest": {
      "delete": {
        "operationId": "undoLatestSelection",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoPhotoSelectionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Undo this member’s most recent private decision",
        "tags": [
          "curation"
        ]
      }
    },
    "/v1/events/{eventId}/download-manifest": {
      "get": {
        "operationId": "getDownloadManifest",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "KEPT"
              ]
            }
          },
          {
            "name": "photoId",
            "required": false,
            "in": "query",
            "description": "Limit the manifest to one explicitly requested photo, including the viewer’s own contribution.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 100,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadManifestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Get paginated short-lived signed original URLs",
        "tags": [
          "curation"
        ]
      }
    },
...
  },
  "info": {
    "title": "PhotoDome API",
    "description": "Backend API for the PhotoDome iPhone app.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "eventCapability": {
        "scheme": "bearer",
        "bearerFormat": "event capability",
        "type": "http",
        "description": "Opaque event-scoped host or guest capability."
      }
    },
    "schemas": {
...
      "EventViewerDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "HOST",
              "GUEST"
            ]
          }
        },
        "required": [
          "memberId",
          "role"
        ]
      },
      "EventSnapshotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "James's birthday"
          },
          "hostDisplayName": {
            "type": "string",
            "example": "James"
          },
          "locationLabel": {
            "type": "string",
            "nullable": true,
            "example": "Shibuya"
          },
          "state": {
            "type": "string",
            "enum": [
              "LIVE",
              "ENDED",
              "EXPIRING"
            ]
          },
          "memberCount": {
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "example": 2
          },
          "readyPhotoCount": {
            "type": "number",
            "minimum": 0,
            "example": 24
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-25T00:00:00.000Z"
          },
          "endedAt": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "nullable": true
          },
          "uploadsRestrictedAt": {
            "type": "string",
            "nullable": true
          },
          "viewer": {
            "$ref": "#/components/schemas/EventViewerDto"
          }
        },
        "required": [
          "id",
          "name",
          "hostDisplayName",
          "state",
          "memberCount",
          "readyPhotoCount",
          "createdAt",
          "viewer"
        ]
      },
...
      "CreateHostTransferResponseDto": {
        "type": "object",
        "properties": {
          "transferToken": {
            "type": "string",
            "example": "pdt_OpaqueOneTimeTransferToken",
            "writeOnly": true
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-07-25T00:10:00.000Z"
          }
        },
        "required": [
          "transferToken",
          "expiresAt"
        ]
      },
...
      "EventMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Taylor"
          },
          "role": {
            "type": "string",
            "enum": [
              "HOST",
              "GUEST"
            ]
          },
          "joinedAt": {
            "type": "string"
          },
          "isViewer": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "displayName",
          "role",
          "joinedAt",
          "isViewer"
        ]
      },
      "UpdateDisplayNameDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Taylor",
            "minLength": 1,
            "maxLength": 50
          }
        },
        "required": [
          "displayName"
        ]
      },
      "RotatedJoinCodeDto": {
        "type": "object",
        "properties": {
          "joinCode": {
            "type": "string",
            "example": "8NPQ6RWT"
          }
        },
        "required": [
          "joinCode"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ReviewPhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "contributorMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "capturedAt": {
            "type": "string",
            "nullable": true
          },
          "readyAt": {
            "type": "string"
          },
          "displayUrl": {
            "type": "string",
            "format": "uri"
          },
          "thumbnailUrl": {
            "type": "string",
            "format": "uri"
          },
          "urlsExpireAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "contributorMemberId",
          "width",
          "height",
          "readyAt",
          "displayUrl",
          "thumbnailUrl",
          "urlsExpireAt"
        ]
      },
      "ReviewPhotoPageDto": {
        "type": "object",
        "properties": {
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewPhotoDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "readyPhotoCount": {
            "type": "number"
          },
          "decidedPhotoCount": {
            "type": "number"
          },
          "keptPhotoCount": {
            "type": "number"
          }
        },
        "required": [
          "photos",
          "readyPhotoCount",
          "decidedPhotoCount",
          "keptPhotoCount"
        ]
      },
      "SetPhotoSelectionDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "KEEP",
              "SKIP"
            ]
          }
        },
        "required": [
          "decision"
        ]
      },
      "PhotoSelectionDto": {
        "type": "object",
        "properties": {
          "photoId": {
            "type": "string",
            "format": "uuid"
          },
          "decision": {
            "type": "string",
            "enum": [
              "KEEP",
              "SKIP"
            ]
          },
          "decidedAt": {
            "type": "string"
          }
        },
        "required": [
          "photoId",
          "decision",
          "decidedAt"
        ]
      },
      "UndoPhotoSelectionDto": {
        "type": "object",
        "properties": {
          "selection": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhotoSelectionDto"
              }
            ]
          }
        }
      },
      "DownloadManifestPhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "contentType": {
            "type": "string"
          },
          "byteSize": {
            "type": "number"
          },
          "capturedAt": {
            "type": "string",
            "nullable": true
          },
          "readyAt": {
            "type": "string"
          },
          "originalUrl": {
            "type": "string",
            "format": "uri"
          },
          "urlExpiresAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "contentType",
          "byteSize",
          "readyAt",
          "originalUrl",
          "urlExpiresAt"
        ]
      },
      "DownloadManifestDto": {
        "type": "object",
        "properties": {
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DownloadManifestPhotoDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "totalPhotoCount": {
            "type": "number"
          }
        },
        "required": [
          "photos",
          "totalPhotoCount"
        ]
      },
...
    }
  }
}